Made changes in the constructor to allow the resizing of the window.  Intended behaviour is to allow the window to expand in both width and height but not allow the window to shrink smaller than the existing default size.
Other changes to existing components within the frame ensure they resize accordingly.

I'm not completely happy with the change however, the window can still be shrunk slightly after display, which isn't intended.  This means that the gap at the bottom and the right of the frame aren't maintained.  While not a show-stopper, it is slightly annoying!